# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "BitBox API",
    "description": "Rust API for 360 Project notation online",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "paths": {
...
    "/marks/evaluate/group/{group_id}": {
      "post": {
        "tags": [
          "Evaluation"
        ],
        "summary": "Evaluate group",
        "description": "This endpoint allows student to evaluate his group.",
        "operationId": "evaluate_group_route",
        "parameters": [
          {
            "name": "group_id",
            "in": "path",
            "description": "The group id to evaluate",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "The graded students of a group",
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/GradedStudentPostModel"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The group was evaluated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "default": null,
                  "nullable": true
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                },
                "examples": {
                  "InvalidToken": {
                    "value": "Error"
                  },
                  "NoToken": {
                    "value": "Token not provided"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundError"
                },
                "example": "NotFoundError"
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalError"
                },
                "example": "InternalError"
              }
            }
          }
        }
      }
    },
...
    "/groups/project/{id}": {
      "get": {
        "tags": [
          "Groups"
        ],
        "summary": "Get all the groups and the students from a project",
        "description": "This endpoint allows you to get all the groups and the corresponding students from a project in the database.",
        "operationId": "get_groups_and_students_from_project_id_route",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The project id to get the groups from",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "All the returned groups",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectGroupsGetModel"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                },
                "examples": {
                  "InvalidToken": {
                    "value": "Error"
                  },
                  "NoToken": {
                    "value": "Token not provided"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundError"
                },
                "example": "NotFoundError"
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalError"
                },
                "example": "InternalError"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Groups"
        ],
        "summary": "Create a group and add it to an existing project",
        "description": "This endpoint allows you to create a group and add it to an existing project in the database.",
        "operationId": "create_group_route",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The project id to add the group to",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "The group to create",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewGroupPostModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "The group has been created",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequestError"
                },
                "examples": {
                  "InvalidName": {
                    "value": "Invalid name"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                },
                "examples": {
                  "InvalidToken": {
                    "value": "Error"
                  },
                  "NoToken": {
                    "value": "Token not provided"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Project Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundError"
                },
                "example": "NotFoundError"
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalError"
                },
                "example": "InternalError"
              }
            }
          }
        }
      }
    },
    "/groups/project/{id}/students": {
      "get": {
        "tags": [
          "Groups"
        ],
        "summary": "Get the students which aren't in a group for a project",
        "description": "This endpoint returns all the students which aren't in a group for a chosen project",
        "operationId": "get_students_without_group_route",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The project id to get the students from",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "All the returned students",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Student"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                },
                "examples": {
                  "InvalidToken": {
                    "value": "Error"
                  },
                  "NoToken": {
                    "value": "Token not provided"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundError"
                },
                "example": "NotFoundError"
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalError"
                },
                "example": "InternalError"
              }
            }
          }
        }
      }
    },
    "/groups/{group_id}/student/{student_id}": {
      "get": {
        "tags": [
          "Groups"
        ],
        "summary": "Returns the details of the 360 mark for a student in a group",
        "description": "This endpoint returns the details of the 360 mark for a student in a group containing all the marks and comments given by the other students of the group.",
        "operationId": "get_group_student_mark_details_route",
        "parameters": [
          {
            "name": "group_id",
            "in": "path",
            "description": "The group id to get the student from",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "student_id",
            "in": "path",
            "description": "The student id to get the details from",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The student details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentGroupMarkDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                },
                "examples": {
                  "InvalidToken": {
                    "value": "Error"
                  },
                  "NoToken": {
                    "value": "Token not provided"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Group or Student Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundError"
                },
                "example": "NotFoundError"
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalError"
                },
                "example": "InternalError"
              }
            }
          }
        }
      }
    },
    "/groups/{group_id}/students": {
      "post": {
        "tags": [
          "Groups"
        ],
        "summary": "Assigns students to a group",
        "description": "This endpoint allows you to assign students to a group in the database.",
        "operationId": "assign_students_to_group_route",
        "parameters": [
          {
            "name": "group_id",
            "in": "path",
            "description": "The group id to assign the students to",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "The students ids to assign",
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The students have been assigned to the group"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                },
                "examples": {
                  "InvalidToken": {
                    "value": "Error"
                  },
                  "NoToken": {
                    "value": "Token not provided"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Group Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundError"
                },
                "example": "NotFoundError"
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalError"
                },
                "example": "InternalError"
              }
            }
          }
        }
      }
    },
    "/groups/{id}": {
      "put": {
        "tags": [
          "Groups"
        ],
        "summary": "Update a group",
        "description": "This endpoint allows you to update name or mark of a group in the database.",
        "operationId": "update_group_route",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The group id to update",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "The group to update",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatedGroupPutModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The group has been updated"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequestError"
                },
                "examples": {
                  "InvalidName": {
                    "value": "Invalid name"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                },
                "examples": {
                  "InvalidToken": {
                    "value": "Error"
                  },
                  "NoToken": {
                    "value": "Token not provided"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Group Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundError"
                },
                "example": "NotFoundError"
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalError"
                },
                "example": "InternalError"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Groups"
        ],
        "summary": "Delete a group",
        "description": "This endpoint allows you to delete a group in the database.",
        "operationId": "delete_group_route",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The group id to delete",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The group has been deleted"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                },
                "examples": {
                  "InvalidToken": {
                    "value": "Error"
                  },
                  "NoToken": {
                    "value": "Token not provided"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Group Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundError"
                },
                "example": "NotFoundError"
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalError"
                },
                "example": "InternalError"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BadRequestError": {
        "type": "object",
        "required": [
          "request"
        ],
        "properties": {
          "request": {
            "type": "string"
          }
        }
      },
...
      "GradedStudentPostModel": {
        "type": "object",
        "required": [
          "student_id",
          "mark"
        ],
        "properties": {
          "comment": {
            "type": "string",
            "nullable": true
          },
          "mark": {
            "type": "number",
            "format": "double"
          },
          "student_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "InternalError": {
        "default": null,
        "nullable": true
      },
...
      "NewGroupPostModel": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
...
      "NotFoundError": {
        "type": "object",
        "required": [
          "resource"
        ],
        "properties": {
          "resource": {
            "type": "string"
          }
        }
      },
...
      "ProjectGroupsGetModel": {
        "type": "object",
        "required": [
          "groups"
        ],
        "properties": {
          "groups": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProjectGroup"
            }
          }
        }
      },
...
      "Student": {
        "type": "object",
        "required": [
          "id",
          "name",
          "surname",
          "email"
        ],
        "properties": {
          "email": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "surname": {
            "type": "string"
          }
        }
      },
...
      "UnauthorizedError": {
        "default": null,
        "nullable": true
      },
...
      "UpdatedGroupPutModel": {
        "type": "object",
        "properties": {
          "mark": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
    },
...
    "securitySchemes": {
      "api_key": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "api_key": [
        "read",
        "write"
      ]
    }
  ]
}